+++ tests/system_tests/test_hw_cpx_e.py
@@ -44,12 +44,12 @@ def test_module_count(test_cpxe):
 
 
 def test_fault_detection(test_cpxe):
-    response = test_cpxe.fault_detection()
+    response = test_cpxe.read_fault_detection()
     assert len(response) == 24
 
 
 def test_status_register(test_cpxe):
-    response = test_cpxe.status_register()
+    response = test_cpxe.read_status_register()
     assert response == (False, False)